Changed file locations and re organized files by type
